When skipping bound subtrees match subtrees not equality
Previusly I was comparing bound subtrees by equality, this is wrong
because when a bound subtree was missing from an ancestor the whole
subtree starting at the ancestor was marked for upstream removal on
push. This is wrong because it removed the content of the bound subtree
while leaving the binding configuration active.

This fixes the issue by skipping contained bound subtrees, not expecting
the subtree to be bound identically.

Also added debugging tools for printing file trees to the terminal.