Implement binding creation when files are present locally
Righ now file download isn't implemented so when adding a binding
present upstream but not locally we bailed out. It's possible the files
are already present locally if they are part of an already bound
subtree, in this case we can create the binding without downloading any
files.

Fixed crash when removing directories and updating the remote index
after we receive the update of the removal of the parent.