Support creating bindings for recently downloaded directories
We don't support merging a local directory with one in the remote, but
if the local copy was just downloaded from the remote and it's expected
to be identical, we now allow creating a binding. For this the user
passes the --downloaded flag, signifying that they just downloaded the
local version.

Additionally, I've added a download_and_bind snip that uses rclone to
download a directoy and immediately binds it.